Fixed #25978 -- Deprecated shorcuts.render_to_response().
