+++ docs/topics/http/shortcuts.txt
@@ -86,9 +86,11 @@ This example is equivalent to::
 
 .. function:: render_to_response(template_name, context=None, content_type=None, status=None, using=None)
 
+   .. deprecated:: 2.0
+
    This function preceded the introduction of :func:`render` and works
    similarly except that it doesn't make the ``request`` available in the
-   response. It's not recommended and is likely to be deprecated in the future.
+   response.
 
 ``redirect()``
 ==============
